apps/workers/requirements.txt
fastapi
uvicorn
bullmq
redis
chromadb
pymupdf
langchain
langchain-text-splitters
openai
python-dotenv
asyncpg
